refactor: Move general types to a common file
