+++ app/charts/shared/use-chart-state.tsx
@@ -13,8 +13,8 @@ import { ScatterplotState } from "@/charts/scatterplot/scatterplot-state";
 import { TableChartState } from "@/charts/table/table-state";
 import { ChartFields, InteractiveFiltersConfig } from "@/configurator";
 import { Observation } from "@/domain/data";
+import { Has } from "@/domain/types";
 import { DimensionMetaDataFragment } from "@/graphql/query-hooks";
-import { Has } from "@/lib/has";
 
 export interface ChartProps {
   data: Observation[];
