--- app/lib/has.ts
@@ -1,5 +1,0 @@
-export type Has<T extends unknown, R extends string> = T extends {
-  [k in R]: any;
-}
-  ? T
-  : never;
