+++ app/pages/__test/index.tsx
@@ -2,8 +2,8 @@ import { groupBy } from "lodash";
 import { GetStaticProps, NextPage } from "next";
 import Link from "next/link";
 
+import { PromiseValue } from "@/domain/types";
 import { loadFixtureConfigs } from "@/test/utils";
-import { PromiseValue } from "@/utils/promise";
 
 type PageProps = {
   configs: PromiseValue<ReturnType<typeof loadFixtureConfigs>>;
