--- app/utils/promise.ts
@@ -1,1 +1,0 @@
-export type PromiseValue<T> = T extends Promise<infer S> ? S : T;
